動作確認用コンポーネントの追加と、読み込み
